splitted example and library
