+++ example/ecore-config.ts
