+++ example/etype.renderer.ts
