+++ example/index.html
@@ -37,5 +37,5 @@
 </nav>
 <div id="editor"></div>
 </body>
-<script src="bundle_ecore.js"></script>
+<script src="bundle.js"></script>
 </html>
