--- graph-editor/che.integration.js
@@ -1,6 +1,0 @@
-function updateEcoreEditorData(element, data) {
-    console.log("Initializing ECore editor...")
-    var ecore = document.createElement('jsonforms-ecore');
-    ecore.data = data;
-    element.appendChild(ecore);
-};
