--- graph-editor/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>JSONGraph Ecore</title>
-  <base href="/">
-
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <script src="assets/native-shim.js"></script>
-  <link rel="stylesheet" href="assets/Roboto-Regular.ttf">
-  <link rel="stylesheet" href="assets/jsonforms-example.css">
-  <link rel="stylesheet" href="assets/jsongraph-example.css">
-  <link rel="stylesheet" href="assets/joint.min.css">
-  <link rel="stylesheet" href="example.css">
-</head>
-<body>
-  <button id="saveData">Push Data</button>
-</body>
-<script src="assets/bundle_graph.js"></script>
-</html>
