+++ tsconfig.json
@@ -15,6 +15,6 @@
     "node_modules"
   ],
   "files": [
-    "./src/index.ts"
+    "./src/jsoneditor.ts"
   ]
 }
